+++ node/utils/glovo.ts
@@ -1,271 +1,9 @@
-/* eslint-disable @typescript-eslint/naming-convention */
-import type {
-  OrderFormItem,
-  SimulationPayload,
-  PayloadItem,
-  Checkout,
-} from '@vtex/clients'
-
+import { HANDLING, ACCEPTED, INVOICED, READY_FOR_PICKUP } from '../constants'
 import {
-  ACCEPTED,
-  HANDLING,
-  HOME,
-  INVOICED,
-  READY_FOR_PICKUP,
-  RESIDENTIAL,
-} from './constants'
-
-export const isSkuAvailable = (item: OrderFormItem | undefined): boolean => {
-  if (!item) {
-    return false
-  }
-
-  return item.availability === 'available'
-}
-
-export const createSimulationItem = ({
-  id,
-  quantity,
-}: {
-  id: string
-  quantity: number
-}): PayloadItem => {
-  return {
-    id,
-    quantity,
-    seller: '1',
-  }
-}
-
-export const createSimulationPayload = ({
-  items,
-  postalCode,
-  country,
-  storeId,
-  salesChannel,
-}: CreateSimulationArgs): [SimulationPayload, string] => {
-  const simulationPayload = {
-    items,
-    postalCode,
-    country,
-  }
-
-  const queryString = `?storeId=${storeId}&sc=${salesChannel}`
-
-  return [simulationPayload, queryString]
-}
-
-const simulateItem = async (
-  IdSku: string,
-  store: StoreInfo,
-  checkout: Checkout
-): Promise<{ price: number; available: boolean }> => {
-  const { storeId, salesChannel, postalCode, country } = store
-  const simulationItem = createSimulationItem({ id: IdSku, quantity: 1 })
-  const simulationPayload = createSimulationPayload({
-    items: [simulationItem],
-    storeId,
-    salesChannel,
-    postalCode,
-    country,
-  })
-
-  const simulation = await checkout.simulation(...simulationPayload)
-
-  const {
-    items: [item],
-  } = simulation
-
-  let itemInfo = {
-    price: 0,
-    available: false,
-  }
-
-  if (isSkuAvailable(item)) {
-    const { price, listPrice, unitMultiplier } = item
-
-    itemInfo = {
-      price: (Math.max(price, listPrice) * unitMultiplier) / 100,
-      available: true,
-    }
-  }
-
-  return itemInfo
-}
-
-export const getStoreInfoFormGlovoStoreId = (
-  id: string,
-  stores: StoreInfo[]
-): StoreInfo | undefined =>
-  stores.find(({ glovoStoreId }) => glovoStoreId === id)
-
-export const getStoreInfoFromStoreId = (
-  id: string,
-  stores: StoreInfo[]
-): StoreInfo | undefined => stores.find(({ storeId }) => storeId === id)
-
-export const convertGlovoProductToItems = (
-  glovoProducts: GlovoProduct[] = []
-): PayloadItem[] => {
-  const items = []
-  const attributesCollection = []
-
-  for (const product of glovoProducts) {
-    const { id, quantity, attributes = [] } = product
-    const item = createSimulationItem({ id, quantity })
-
-    items.push(item)
-    attributesCollection.push(...attributes)
-  }
-
-  for (const attribute of attributesCollection) {
-    const { id, quantity } = attribute
-    const item = createSimulationItem({ id, quantity })
-
-    items.push(item)
-  }
-
-  return items
-}
-
-export const convertGlovoProductsToCompare = (
-  glovoProducts: GlovoProduct[]
-) => {
-  const items = []
-  const attributesCollection = []
-
-  for (const product of glovoProducts) {
-    const { id, quantity, attributes = [], purchased_product_id } = product
-    const item = {
-      id,
-      quantity,
-      purchased_product_id,
-    }
-
-    items.push(item)
-    attributesCollection.push(...attributes)
-  }
-
-  for (const attribute of attributesCollection) {
-    if (attribute.purchased_product_id) {
-      const { id, quantity } = attribute
-
-      const updatedItems: GlovoProductAttributes[] = items.map((item) => {
-        if (item.id === id) {
-          item.quantity += quantity
-        }
-
-        return item
-      })
-
-      items.push(...updatedItems)
-    }
-  }
-
-  return items
-}
-
-export const createVtexOrderData = (
-  glovoOrder: GlovoOrder,
-  orderSimulation: any,
-  clientProfileData: ClientProfileData
-): MarketplaceOrder => {
-  const { order_id, estimated_total_price } = glovoOrder
-  const { items, pickupPoints, postalCode, logisticsInfo } = orderSimulation
-  const {
-    email,
-    firstName,
-    lastName,
-    documentType,
-    document,
-    phone,
-    corporateName,
-  } = clientProfileData
-
-  /** Re-Index items array */
-  let counter = 0
-  const updatedItems = items.reduce((acc: any, item: any) => {
-    if (isSkuAvailable(item)) {
-      const updatedItem = {
-        ...item,
-        itemIndex: 0,
-      }
-
-      counter++
-
-      return [...acc, updatedItem]
-    }
-
-    return acc
-  }, [])
-
-  /** Re-Index logisticsInfo array */
-  counter = 0
-  const updatedLogisticsInfo = logisticsInfo.reduce(
-    (acc: any[], item: any, index: number) => {
-      if (items[index].availability === 'available') {
-        const updatedItem = {
-          itemIndex: counter,
-          selectedSla: item.slas[0].id,
-          selectedDeliveryChannel: item.slas[0].deliveryChannel,
-          name: item.slas[0].name,
-          lockTTL: '1d',
-          deliveryWindow: item.slas[0].availableDeliveryWindows[0],
-          shippingEstimate: item.slas[0].shippingEstimate,
-          price: item.slas[0].price,
-        }
-
-        counter++
-
-        return [...acc, updatedItem]
-      }
-
-      return acc
-    },
-    []
-  )
-
-  return {
-    marketplaceOrderId: order_id,
-    marketplaceServicesEndpoint: 'https://api.glovoapp.com/',
-    isCreatedAsync: true,
-    marketplacePaymentValue: estimated_total_price,
-    items: updatedItems,
-    clientProfileData: {
-      email,
-      firstName,
-      lastName,
-      documentType,
-      document,
-      phone,
-      corporateName,
-      tradeName: null,
-      corporateDocument: null,
-      stateInscription: null,
-      corporatePhone: null,
-      isCorporate: false,
-      userProfileId: null,
-    },
-    shippingData: {
-      address: {
-        addressType: RESIDENTIAL,
-        receiverName: `${firstName} ${lastName}`,
-        addressId: HOME,
-        postalCode,
-        city: pickupPoints[0].address.city,
-        state: pickupPoints[0].address.state,
-        country: pickupPoints[0].address.country,
-        street: pickupPoints[0].address.street,
-        number: null,
-        neighborhood: null,
-        complement: null,
-        reference: null,
-        geoCoordinates: [],
-      },
-      logisticsInfo: updatedLogisticsInfo,
-    },
-  }
-}
+  createSimulationItem,
+  createSimulationPayload,
+  simulateItem,
+} from './simulation'
 
 export const setGlovoStatus = (state: string) => {
   if (state === HANDLING) return ACCEPTED
